cache.off {pander}	R Documentation

Toggle cache

Description

This function is just a wrapper around evalsOptions to switch pander's cache on or off easily, which might be handy in some brew documents to prevent repetitive strain injury :)

Usage

cache.on()

cache.off()

[Package pander version 0.6.3 Index]