coef_mat {pander}	R Documentation

Calculate coef matrix for models from rms package Forked from prModFit from rms

Description

Calculate coef matrix for models from rms package Forked from prModFit from rms

Usage

coef_mat(obj, coefs)

Arguments

obj
object list

coefs
numeric value if to print only the first n regression coefficients in the model.

Value

coeficients matrix

[Package pander version 0.6.3 Index]