pander.Arima {pander}	R Documentation

Prints an arima object from stats package in Pandoc's markdown.

Description

Prints an arima object from stats package in Pandoc's markdown.

Usage

## S3 method for class 'Arima'
pander(x, digits = panderOptions("digits"), se = TRUE,
  ...)

Arguments

x
an arima object

digits
number of digits of precision

se
if to include standard error in coefficients table (default TRUE)

...
optional parameters passed to raw pandoc.table function

[Package pander version 0.6.3 Index]