pander.cph {pander}	R Documentation

Prints an cph object from rms package in Pandoc's markdown.

Description

Prints an cph object from rms package in Pandoc's markdown.

Usage

## S3 method for class 'cph'
pander(x, table = TRUE, conf.int = FALSE, coefs = TRUE,
  ...)

Arguments

x
an cph object

table
if to print event frequency statistics. default(TRUE)

conf.int
set to e.g. .95 to print 0.95 confidence intervals on simple hazard ratios (which are usually meaningless as one-unit changes are seldom relevant and most models contain multiple terms per predictor)

coefs
if to the table of model coefficients, standard errors, etc. default(TRUE)

...
optional parameters passed to raw pandoc.table function

[Package pander version 0.6.3 Index]