pander.ets {pander}	R Documentation

Prints an ets object from forecast package in Pandoc's markdown.

Description

Prints an ets object from forecast package in Pandoc's markdown.

Usage

## S3 method for class 'ets'
pander(x, digits = panderOptions("digits"), ...)

Arguments

x
an ets object

digits
number of digits of precision

...
optional parameters passed to raw pandoc.table function

[Package pander version 0.6.5 Index]