pander.function {pander}	R Documentation

Pander method for function class

Description

Prints an function object in Pandoc's markdown.

Usage

## S3 method for class ''function''
pander(x, add.name = FALSE, verbatim = TRUE, syntax.highlighting = FALSE, ...)

Arguments

x
an function object

add.name
(defaut:FALSE) if to add function name to output or just to print a body

verbatim
(defaut:TRUE) if to add tabulation, so pandoc conversion will rander it properly

syntax.highlighting
(defaut:FALSE) if to add hyghlighting tag for R syntax

...
ignored parameters

[Package pander version 0.6.5 Index]