pander.irts {pander}	R Documentation

Prints an irts object from tseries package in Pandoc's markdown.

Description

Prints an irts object from tseries package in Pandoc's markdown.

Usage

## S3 method for class 'irts'
pander(x, caption = attr(x, "caption"), format = panderOptions("date"), ...)

Arguments

x
an irts object

caption
caption (string) to be shown under the table

format
string passed to format when printing dates (POSIXct or POSIXt)

...
optional parameters passed to raw pandoc.table function

[Package pander version 0.6.5 Index]