pander.ols {pander}	R Documentation

Prints an ols object from rms package in Pandoc's markdown.

Description

Prints an ols object from rms package in Pandoc's markdown.

Usage

## S3 method for class 'ols'
pander(x, long = FALSE, coefs = TRUE,
  digits = panderOptions("digits"), round = panderOptions("round"),
  ...)

Arguments

x
an ols object

long
if to print the correlation matrix of parameter estimates. default(FALSE)

coefs
if to the table of model coefficients, standard errors, etc. default(TRUE)

digits
passed to format. Can be a vector specifying values for each column (has to be the same length as number of columns).

round
passed to round. Can be a vector specifying values for each column (has to be the same length as number of columns). Values for non-numeric columns will be disregarded.

...
optional parameters passed to raw pandoc.table function

[Package pander version 0.6.3 Index]