pander.orm {pander}	R Documentation

Prints an orm object from rms package in Pandoc's markdown.

Description

Prints an orm object from rms package in Pandoc's markdown.

Usage

## S3 method for class 'orm'
pander(x, coefs = TRUE, intercepts = x$non.slopes < 10,
  ...)

Arguments

x
an orm object

coefs
if to the table of model coefficients, standard errors, etc. default(TRUE)

intercepts
if to print intercepts, by default, intercepts are only printed if there are fewer than 10 of them

...
optional parameters passed to raw pandoc.table function

[Package pander version 0.6.3 Index]