pander.summary.glm {pander}	R Documentation

Pander method for summary.glm class

Description

Prints a summary.glm object in Pandoc's markdown.

Usage

## S3 method for class 'summary.glm'
pander(
  x,
  caption = attr(x, "caption"),
  covariate.labels,
  omit,
  summary = TRUE,
  ...
)

Arguments

x
an summary.glm object

caption
caption (string) to be shown under the table

covariate.labels
vector to replace covariate lables in the table

omit
vector of variable to omit for priting in resulting table

summary
(defaut:TRUE) if used for summary.lm or lm

...
optional parameters passed to special methods and/or raw pandoc.* functions

Value

By default this function outputs (see: cat) the result. If you would want to catch the result instead, then call the function ending in .return.

[Package pander version 0.6.5 Index]