pander.summary.lm {pander}	R Documentation

Pander method for summary.lm class

Description

Prints a summary.lm object in Pandoc's markdown.

Usage

## S3 method for class 'summary.lm'
pander(x, caption = attr(x, "caption"),
  covariate.labels, omit, summary = TRUE,
  add.significance.stars = FALSE, move.intercept = FALSE, ...)

Arguments

x
an summary.lm object

caption
caption (string) to be shown under the table

covariate.labels
vector to replace covariate lables in the table

omit
vector of variable to omit for priting in resulting table

summary
(defaut:TRUE) if used for summary.lm or lm

add.significance.stars
if significance stars should be shown for P value

move.intercept
by default, the Intercept is the first coefficient in the table, which can be moved to the bottom of the table

...
optional parameters passed to special methods and/or raw pandoc.* functions

Value

By default this function outputs (see: cat) the result. If you would want to catch the result instead, then call the function ending in .return.

[Package pander version 0.6.3 Index]