pander.summary.survreg {pander}	R Documentation

Prints an survreg object from survival package in Pandoc's markdown.

Description

Prints an survreg object from survival package in Pandoc's markdown.

Usage

## S3 method for class 'summary.survreg'
pander(x, summary = TRUE,
  digits = panderOptions("digits"), round = panderOptions("round"),
  keep.trailing.zeros = panderOptions("keep.trailing.zeros"), ...)

Arguments

x
an survreg object

summary
if summary should be printed

digits
number of digits of precision passed to format

round
number of rounding digits passed to round

keep.trailing.zeros
to show or remove trailing zeros in numbers on a column basis width

...
optional parameters passed to raw pandoc.table function

[Package pander version 0.6.3 Index]