pander.survfit {pander}	R Documentation

Pander method for survfit class

Description

Prints an survfit object in Pandoc's markdown.

Usage

## S3 method for class 'survfit'
pander(x, caption = attr(x, "caption"), scale = 1,
  print.rmean = getOption("survfit.print.rmean"),
  rmean = getOption("survfit.rmean"), ...)

Arguments

x
the result of a call to the survfit function.

caption
caption (string) to be shown under the table

scale
a numeric value to rescale the survival time, e.g., if the input data to survfit were in days, scale=365 would scale the printout to years.

print.rmean, rmean
options for computation and display of the restricted mean

...
optional parameters passed to raw pandoc.table function

[Package pander version 0.6.3 Index]