panderOptions {pander}	R Documentation

Querying/setting pander option

Description

To list all pander options, just run this function without any parameters provided. To query only one value, pass the first parameter. To set that, use the value parameter too.

Usage

panderOptions(o, value)

Arguments

o
option name (string). See below.

value
value to assign (optional)

Details

The following pander options are available:

digits: numeric (default: 2) passed to format. Can be a vector specifying values for each column (has to be the same length as number of columns). Values for non-numeric columns will be disregarded.

decimal.mark: string (default: .) passed to format

formula.caption.prefix: string (default: 'Formula: ') passed to pandoc.formula to be used as caption prefix. Be sure about what you are doing if changing to other than 'Formula: ' or ':'.

big.mark: string (default: ”) passed to format.

round: numeric (default: Inf) passed to round. Can be a vector specifying values for each column (has to be the same length as number of columns). Values for non-numeric columns will be disregarded.

keep.trailing.zeros: boolean (default: FALSE) to show or remove trailing zeros in numbers

keep.line.breaks: boolean (default: FALSE) to keep or remove line breaks from cells in a table

missing: string (default: NA) to replace missing values in vectors, tables etc.

date: string (default: '%Y/%m/%d %X') passed to format when printing dates (POSIXct or POSIXt)

header.style: 'atx' or 'setext' passed to pandoc.header

list.style: 'bullet', 'ordered' or 'roman' passed to pandoc.list. Please not that this has no effect on pander methods.

table.style: 'multiline', 'grid', 'simple' or 'rmarkdown' passed to pandoc.table

table.emphasize.rownames: boolean (default: TRUE) if row names should be highlighted

table.split.table: numeric passed to pandoc.table and also affects pander methods. This option tells pander where to split too wide tables. The default value (80) suggests the conventional number of characters used in a line, feel free to change (e.g. to Inf to disable this feature) if you are not using a VT100 terminal any more :)

table.split.cells: numeric or numeric vector (default: 30) passed to pandoc.table and also affects pander methods. This option tells pander where to split too wide cells with line breaks. Numeric vector specifies values for cells separately. Set Inf to disable.

table.caption.prefix: string (default: 'Table: ') passed to pandoc.table to be used as caption prefix. Be sure about what you are doing if changing to other than 'Table: ' or ':'.

table.continues: string (default: 'Table continues below') passed to pandoc.table to be used as caption for long (split) without a use defined caption

table.continues.affix: string (default: '(continued below)') passed to pandoc.table to be used as an affix concatenated to the user defined caption for long (split) tables

table.alignment.default: string (default: centre) that defines the default alignment of cells. Can be left, right or centre that latter can be also spelled as center.

table.alignment.rownames: string (default: centre) that defines the alignment of rownames in tables. Can be left, right or centre that latter can be also spelled as center.

use.hyphening: boolean (default: FALSE) if try to use hyphening when splitting large cells according to table.split.cells. Requires sylly.

evals.messages: boolean (default: TRUE) passed to evals' pander method specifying if messages should be rendered

p.wrap: a string (default: '_') to wrap vector elements passed to p function

p.sep: a string (default: ', ') with the main separator passed to p function

p.copula: a string (default: ' and ') with ending separator passed to p function

plain.ascii: boolean (default: FALSE) to define if output should be in plain ascii or not

graph.nomargin: boolean (default: TRUE) if trying to keep plots' margins at minimal

graph.fontfamily: string (default: 'sans') specifying the font family to be used in images. Please note, that using a custom font on Windows requires grDevices:::windowsFonts first.

graph.fontcolor: string (default: 'black') specifying the default font color

graph.fontsize: numeric (default: 12) specifying the base font size in pixels. Main title is rendered with 1.2 and labels with 0.8 multiplier.

graph.grid: boolean (default: TRUE) if a grid should be added to the plot

graph.grid.minor: boolean (default: TRUE) if a miner grid should be also rendered

graph.grid.color: string (default: 'grey') specifying the color of the rendered grid

graph.grid.lty: string (default: 'dashed') specifying the line type of grid

graph.boxes: boolean (default: FALSE) if to render a border around of plot (and e.g. around strip)

graph.legend.position: string (default: 'right') specifying the position of the legend: 'top', 'right', 'bottom' or 'left'

graph.background: string (default: 'white') specifying the plots main background's color

graph.panel.background: string (default: 'transparent') specifying the plot's main panel background. Please note, that this option is not supported with base graphics.

graph.colors: character vector of default color palette (defaults to a colorblind theme: http://jfly.iam.u-tokyo.ac.jp/color/). Please note that this update work with base plots by appending the col argument to the call if not set.

graph.color.rnd: boolean (default: FALSE) specifying if the palette should be reordered randomly before rendering each plot to get colorful images

graph.axis.angle: numeric (default: 1) specifying the angle of axes' labels. The available options are based on par(les) and sets if the labels should be:

1: parallel to the axis,

2: horizontal,

3: perpendicular to the axis or

4: vertical.

graph.symbol: numeric (default: 1) specifying a symbol (see the pch parameter of par)

knitr.auto.asis: boolean (default: TRUE) if the results of pander should be considered as 'asis' in knitr. Equals to specifying results='asis' in the R chunk, so thus there is no need to do so if set to TRUE.

pandoc.binary: full path of pandoc's binary. By default, pandoc is in the path.

See Also

evalsOptions

Examples

## Not run: 
panderOptions()
panderOptions('digits')
panderOptions('digits', 5)

## End(Not run)

[Package pander version 0.6.3 Index]