pandoc.date.return {pander}	R Documentation

Dates

Description

Pandoc's mardown date.

Usage

pandoc.date.return(x, inline = TRUE, simplified = FALSE, ...)

Arguments

x
date or vector of dates

inline
if to render vector of dates as inline paragraph or not (as list)

simplified
if just add date formatting to vector of dates

...
extra arguments passed by from parent call, disregarded

Value

By default this function outputs (see: cat) the result. If you would want to catch the result instead, then call the function ending in .return.

Examples

pandoc.date(Sys.Date())
pandoc.date(Sys.Date() - 1:10)
pandoc.date(Sys.Date() - 1:10, inline = FALSE)

[Package pander version 0.6.3 Index]