pandoc.formula.return {pander}	R Documentation

Formulas

Description

Pandoc's mardown formula.

Usage

pandoc.formula.return(
  x,
  text = NULL,
  max.width = 80,
  caption,
  add.line.breaks = FALSE,
  ...
)

Arguments

x
formula

text
text to be written before result in the same line. Typically used by calls from other functions in the package

max.width
maximum width in characters per line

caption
caption (string) to be shown under the formula

add.line.breaks
if to add 2 line breaks after formula

...
extra arguments passed by from parent call, disregarded

Value

By default this function outputs (see: cat) the result. If you would want to catch the result instead, then call the function ending in .return.

Examples

pandoc.formula(y ~ x)
pandoc.formula(formula(paste('y ~ ', paste0('x', 1:12, collapse = ' + '))))

[Package pander version 0.6.5 Index]