pandoc.image.return {pander}	R Documentation

Create pandoc image tags

Description

Creates a Pandoc's markdown format image hyperlink.

Usage

pandoc.image.return(img, caption = storage$caption)

Arguments

img
image path

caption
text

Value

By default this function outputs (see: cat) the result. If you would want to catch the result instead, then call the function ending in .return.

Note

The caption text is read from an internal buffer which defaults to NULL. To update that, call link{set.caption} before.

References

John MacFarlane (2012): _Pandoc User's Guide_. https://johnmacfarlane.net/pandoc/README.html

See Also

set.caption

Examples

pandoc.image('foo.png')
pandoc.image('foo.png', 'Nice image, huh?')

[Package pander version 0.6.5 Index]