path_to_pandoc {pander}	R Documentation

Find path to the pandoc binary by checking the PATH and the RSTUDIO_PANDOC env vars

Description

Find path to the pandoc binary by checking the PATH and the RSTUDIO_PANDOC env vars

Usage

path_to_pandoc()

Value

file path

[Package pander version 0.6.5 Index]