redrawPlot {pander}	R Documentation

Redraw a recordedplot, grid, trellis, or ggplot2 plot.

Description

This function redraws the plot represented by rec_plot. It can redraw grid/trellis/ggplot2/etc plots, as well as recordedplot objects. For recordedplot objects it acts as a wrapper around replayPlot with memory tweaks to fix native symbol address errors when the recordedplot was loaded from an rda/rds file.

Usage

redrawPlot(rec_plot)

Arguments

rec_plot
the plot object to redraw

References

Thanks to Jeroen Ooms https://stat.ethz.ch/pipermail/r-devel/2012-January/062973.html, JJ Allaire https://github.com/rstudio/rstudio/commit/eb5f6f1db4717132c2ff111f068ffa6e8b2a5f0b, and Gabriel Becker.

See Also

redraw.recordedplot

[Package pander version 0.6.3 Index]