set.alignment {pander}	R Documentation

Sets alignment for tables

Description

This is a helper function to update the alignment (justify parameter in pandoc.table) of the next returning table. Possible values are: centre or center, right, left.

Usage

set.alignment(default = panderOptions("table.alignment.default"),
  row.names = panderOptions("table.alignment.rownames"),
  permanent = FALSE)

Arguments

default
character vector which length equals to one (would be repeated n times) ot n - where n equals to the number of columns in the following table

row.names
string holding the alignment of the (optional) row names

permanent
(default FALSE) if alignment is permanent (for all future tables) or not. It's cleaner to use panderOptions instead.

[Package pander version 0.6.3 Index]