brew {pander}	R Documentation

Patched brew

Description

This is a forked/patched version of 'brew' package made by Jeffrey Horner (c) 2007. See: References about the original version.

Usage

brew(text = NULL, envir = parent.frame())

Arguments

text
character vector

envir
environment

Details

This custom function can do more and also less compared to the original brew package. First of all the internal caching mechanism (and other, from pander package POV needless features) of 'brew' is removed for some extra profits:

multiple R expressions can be passed between <%= ... %> tags,

the text of the file and also the evaluated R objects are (invisibly) returned in a structured list, which can be really useful while post-processing the results of 'brew'.

Value

brewed document to stdout and raw results while evaluating the text in a structured list.

Note

This function should be never called directly (use brew::brew instead) as being a helper function of Pandoc.brew.

References

Jeffrey Horner (2011). _brew: Templating Framework for Report Generation._ https://cran.r-project.org/package=brew

See Also

Pandoc.brew

[Package pander version 0.6.5 Index]