emphasize.rows {pander}	R Documentation

Emphasize rows/columns/cells

Description

Storing indexes of cells to be (strong) emphasized of a tabular data in an internal buffer that can be released and applied by pandoc.table, pander or evals later.

Usage

emphasize.rows(x)

emphasize.cols(x)

emphasize.cells(x)

emphasize.strong.rows(x)

emphasize.strong.cols(x)

emphasize.strong.cells(x)

emphasize.italics.rows(x)

emphasize.italics.cols(x)

emphasize.italics.cells(x)

emphasize.verbatim.rows(x)

emphasize.verbatim.cols(x)

emphasize.verbatim.cells(x)

Arguments

x
vector of row/columns indexes or an array like returned by which(..., arr.ind = TRUE)

Examples

## Not run: 
n <- data.frame(x = c(1,1,1,1,1), y = c(0,1,0,1,0))
emphasize.cols(1)
emphasize.rows(1)
pandoc.table(n)

emphasize.strong.cells(which(n == 1, arr.ind = TRUE))
pander(n)

## End(Not run)

[Package pander version 0.6.3 Index]