get.alignment {pander}	R Documentation

Get alignment

Description

Get alignment from temporary environment, truncating that and applying rownames and other columns alignment to passed df.

Usage

get.alignment(df, remove.obious.rownames = TRUE)

Arguments

df
data.frame

remove.obious.rownames
if this rule should be applied or not

Value

vector of alignment parameters

[Package pander version 0.6.3 Index]