multitable {pander}	R Documentation

Create a multitable used for rendering objects from rms package

Description

When ols/lrm/orm from rms package get rendered, main statistics are group in table of tables. Since pandoc doesn't support row or col-span, we chose to group those statistics in a column each. This function takes care of that

Usage

multitable(v)

Arguments

v
list of vectors/lists to be merge in the table

Value

data.frame in specified format

Examples

pander:::multitable(list(list(a=1, b=2),list(c=3, d=4)))

[Package pander version 0.6.5 Index]