pander.microbenchmark {pander}	R Documentation

Pander method for microbenchmark class

Description

Prints an microbenchmark object in Pandoc's markdown.

Usage

## S3 method for class 'microbenchmark'
pander(x, caption = attr(x, "caption"), expr.labels, unit, ...)

Arguments

x
an microbenchmark object

caption
caption (string) to be shown under the table

expr.labels
expression labels that will replace default ones (similar to rownames, which microbenchmark class table does not have)

unit
units in which values should be printed (for example second, microseconds, etc.). Should be one of ns, us, ms, s, t, hz, khz, mhz, eps, f

...
optional parameters passed to raw pandoc.table function

[Package pander version 0.6.5 Index]