pandoc.horizontal.rule.return {pander}	R Documentation

Create horizontal rule

Description

Creates a Pandoc's markdown format horizontal line with trailing and leading newlines.

Usage

pandoc.horizontal.rule.return()

Value

By default this function outputs (see: cat) the result. If you would want to catch the result instead, then call the function ending in .return.

References

John MacFarlane (2012): _Pandoc User's Guide_. https://johnmacfarlane.net/pandoc/README.html

[Package pander version 0.6.5 Index]