pandoc.title.return {pander}	R Documentation

Create title block

Description

Creates a Pandoc's markdown style title block with optional author, title and date fields.

Usage

pandoc.title.return(author = "", title = "", date = "")

Arguments

author
character vector or semicolon delimited list of authors without line break

title
character vector of lines of title or multiline string with \n separators

date
any string fit in one line

Value

By default this function outputs (see: cat) the result. If you would want to catch the result instead, then call the function ending in .return.

References

John MacFarlane (2012): _Pandoc User's Guide_. https://johnmacfarlane.net/pandoc/README.html

Examples

pandoc.title('Tom', 'Render pandoc in R', '2012-05-16')
pandoc.title(c('Tom', 'Jerry'), 'Render pandoc in R', '2012-05-16')
pandoc.title('Tom; Jerry', 'Render pandoc in R', '2012-05-16')
pandoc.title('Tom; Jerry', c('Render', 'pandoc', 'in R'), '2012-05-16')
pandoc.title('Tom; Jerry', 'Render\n    pandoc \n    in R', '2012-05-16')

## missing fields

pandoc.title('Tom; Jerry', 'Render pandoc in R')
pandoc.title('Tom; Jerry')
pandoc.title(title = 'Render pandoc in R', date= '2012-05-16')

[Package pander version 0.6.5 Index]